Moves getBestMatch to revenuecatui.
